+++ gba/videostructs.h
@@ -22,8 +22,8 @@ namespace Video
     /// @brief Video file / data information
     struct Info : public FileHeader
     {
-        const uint8_t *fileData = nullptr;  // Pointer to file header data
-        const uint8_t *frameData = nullptr; // Pointer to frame data
+        const uint32_t *fileData = nullptr;  // Pointer to file header data
+        const uint32_t *frameData = nullptr; // Pointer to frame data
     } __attribute__((aligned(4), packed));
 
     /// @brief Chunk of compressed data
@@ -37,7 +37,7 @@ namespace Video
     struct Frame
     {
         uint32_t index = std::numeric_limits<uint32_t>::max(); // Frame index in video
-        uint32_t chunkOffset = 0;                              // Offset to chunk data in Info::frameData
+        uint32_t chunkOffset = 0;                              // Offset to start of chunk in Info::frameData
         uint32_t colorMapOffset = 0;                           // Offset to color map in Info::frameData
         uint32_t compressedSize = 0;                           // Size of frame data in chunk (ONLY frame data, not whole chunk)
     } __attribute__((aligned(4), packed));
